% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_Heatmap.R
\name{Create_HeatMap}
\alias{Create_HeatMap}
\title{Create_HeatMap : nteractive heatmap}
\usage{
Create_HeatMap(
  file1,
  file2,
  names_exp = c("AD_vs_Asym", "AD_vs_Cont"),
  cut_num = 0.8
)
}
\arguments{
\item{file1}{First input file in same GEO2R format}

\item{file2}{Second input file for comparasion in the same format}

\item{names_exp}{List of the names of the experiment or comparission written as a list}

\item{cut_num}{Cutoff value for the logFC value}
}
\value{
Retruns a heatmap using the d3heatmap function that provides an interactive heatmap with the dendrogram
}
\description{
This function creates a heatmap based on the input cutoff value provided for the logFC value and provides
an interactive map for the same.
}
