% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_plots.R
\name{Make_plots}
\alias{Make_plots}
\title{Make_plots: Makes histograms and wordclouds for basic analysis}
\usage{
Make_plots(Retrived_Innatedb, t = TRUE)
}
\arguments{
\item{Retrived_Innatedb}{Input information after getting info from the InnateDB suing the Retrive_from_db function}
}
\description{
Make_plots: Makes histograms and wordclouds for basic analysis
}
\seealso{
Retrive_from_db
}
