% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Retrive_from_db.R
\name{Retrive_from_db}
\alias{Retrive_from_db}
\title{Retrive_from_db: Retriving complete information from InnateDB}
\usage{
Retrive_from_db(immune_screen_data, Database = "Innate")
}
\arguments{
\item{immune_screen_data}{Should consist a column of Gene.symbol or name, as seen in GEO2R output file, and is best when Test_Immune function is used previously to reduce the trial number}

\item{Database}{Defaults to Innate, but the RLR database is under development and thus can be used along with Innate}
}
\value{
Retruns a data frame consisting all the important data present in Innate DB
}
\description{
This function retrives all information about the gene from the InnateDB.
}
