% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Screen_pval.R
\name{Screen_pval}
\alias{Screen_pval}
\title{Screen_Pval: Screening Based on P-value}
\usage{
Screen_pval(main_data, remove_empty = TRUE)
}
\arguments{
\item{main_data}{The data file as loaded from GEO2R results using the import function}

\item{remove_empty}{Defaults to TRUE. It removes the rows with no gene symbol present.}
}
\value{
Returns a data frame with the screened gene symbols with other rows intact.
}
\description{
This function removes the genes with p value greater that 0.05. This function also can be used to remove the rows with no gene symbol present.
}
