% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test_Immune.R
\name{Test_Immune}
\alias{Test_Immune}
\title{Test_Immune: Tests the gene under immune gene database InnateDB}
\usage{
Test_Immune(screened_data, get_whole_info = TRUE)
}
\arguments{
\item{screened_data}{The data frame, preferably from the screening function of p value, to serve as an input file along in the same format as output of selected columns in GEO2R analysis}

\item{get_whole_info}{Defaults to TRUE for further analysis, when FALSE, provides only the list of genes}
}
\value{
Returns a dataframe of the immune genes present in the input data
}
\description{
This funciton allows user to filter if the genes belong in the human immune system, as presented in the InnateDB database. It also has an option of output either being just a list of genes or the whole list for further analysis
}
